Add Telepole script
This script can be used to set up a teleport pole that allows players to
teleport back to their group leaders. It tries to be clever about vehicles and
parachuting, but currently does not handle leaders teleporting back to their
groups.

This addition is in preparation for the removal of all the custom JIP stuff
that was added by F3 and Captain Blaffer, which tends to break Zeus.