Make example loadouts vanilla-only
We used to have example loadouts in simple assignGear that used mod weapons and
gear. The problem with that is that sometimes mods go away or equipment gets
renamed. Instead, now use all-vanilla gear based on what the placed units
already have (for NATO, CSAT, AAF) or would have (FIA).