Fix respawn being delayed forever
Reported by Oriskana.

If `respawnDelay` in `description.ext` is increased noticeably from its
dfefault of five seconds (in the report case, to 15s), the respawn would hang
forever in the thrid person "orbit over the player" mode, with no termination
or way out. Why exactly this would happen is uncler --- the expectation would
be of failure/error messages, but not hanging forever.

As a solution, instead of waiting an arbitrary amount of time, run a
slow-looping `waitUntil` with `player` not being NULL as the condition.