assignGear: fix usage for ACE3/CBA loadouts
This change updates the simple assignGear scripts to allow use of both the vanilla format
for loadouts (as exported by the ACE arsenal before the upstream change) as well as the new
ACE3/CBA extended format that is exported by newer versions. More details below.

CBA/ACE changed what is exported from the Arsenal in v3.15.2. Before that version, the
export function in the ACE Arsenal would give you a unit loadout that could be directly
used with Arma's setUnitLoadout function. After that, the exported string is meant to
be used with CBA's CBA_fnc_setLoadout. The difference is that the latter function expects
a wrapped version of the base game loadout. Specifically, a "classic" loadout looks like
this:

```
  [["arifle_MXC_ACO_F","","","",["30Rnd_65x39_caseless_mag",30] ... ,"ItemWatch",""]]
```

whereas the CBA wrapped version of that loadout is

```
 [[["arifle_MXC_ACO_F","","","",["30Rnd_65x39_caseless_mag",30] ... ,"ItemWatch",""]], []]
```

So the easiest way of distinguishing the two is whether or not it starts with two square
brackets ([[ - base game format) or three ([[[ - CBA extended format).