Re-do the logging functions and all their callsites
The old set of logging functions were confusing in what they would do for the
function tag/callsite inference. Remove all of that adn use just the three
original functions (rp, chat, both). The "ehlog" function was a
spur-of-the-moment thing when `scriptNameParent` turned out to not work in many
cases where it should have.